requirements.txt
PySide6 == 6.2.3
numpy
pillow
mysqlclient
pyinstaller
pyqtgraph
